Walk through the valley of namespaces to separate into files :)
