+++ fit/fit.py
@@ -1,168 +1,14 @@
 #!/usr/bin/env python
 
 import sys
-import math
 import numpy
 import scipy.interpolate
-from scipy.stats import linregress
 import matplotlib.pyplot as plt
-from iminuit import Minuit
 import argparse
-from pymongo import MongoClient
-from pymongo import DESCENDING
 
-class Model(object):
-
-    def edgeTransmission(self,x,D):
-        """
-        Returns the transmission fraction (0-1) when the center of the beam with diameter D is
-        a distance x from the edge. The input x can be a numpy array. The units of x and D are
-        arbitrary but must be the same.
-        """
-        r = 2*x/D
-        t = numpy.zeros_like(r)
-        t[r<=-1] = 1.
-        use = r*r<1
-        ruse = r[use]
-        t[use] = (numpy.arccos(ruse)-ruse*numpy.sqrt(1-ruse*ruse))/numpy.pi
-        return t
-    
-    def tabsTransmission(self,x,tabs,D):
-        """
-        Returns the transmission function (0-1) of a sequence of tabs defined by edges at (x1,x2)
-        assuming a beam of diameter D. The input x can be a numpy array. All units are arbitrary but
-        must be the same.
-        """
-        result = numpy.ones_like(x)
-        for (x1,x2) in tabs:
-            result += edgeTransmission(x-x1,D) - edgeTransmission(x-x2,D)
-        return result
-    
-    def motion(self,dt,L,dtheta,T):
-        """
-        Converts times dt (secs) relative to dead center to transverse positions (mm) along the tabs.
-        The inputs are the effective length (mm) of the pendulum L, the angular displacement (deg)
-        of the pendulum's swing relative to dead center, and the effective period T (secs). The
-        input t can be a numpy array.
-        """
-        # Calculate angular frequency in rad/sec
-        omega = 2*numpy.pi/T
-        # Calculate angular speed at dead center in rad/sec
-        thdot = omega*numpy.sqrt(2*(1-numpy.cos(numpy.deg2rad(dtheta))))
-        # Calculate x(t) taking into account two corrections to the simplest x(t) = L*thdot*t model.
-        # First, the marker raises slightly during the swing so its sideways rate of motion dx/dt is
-        # reduced.  Second, the angular speed decreases as the pendulum moves away from dead center.
-        return L*numpy.sin(thdot/omega*numpy.sin(omega*dt))
-    
-    def physicalModel(self,t0,direction,lo,hi,tabs,D=2.,L=1108.,dtheta=4.66,T=2.0,nsamples=1024,adcTick=1664e-7):
-        """
-        t0 = offset of dead center relative to first sample (ADC samples)
-        """
-        # initialize an array of times (secs) relative to the assumed dead center
-        dt = direction*adcTick*(numpy.arange(nsamples) - t0)
-        # convert to transverse positions (mm) relative to the tabs
-        dx = motion(dt,L,dtheta,T)
-        # calculate expected tranmission fraction (0-1) at each transverse position
-        trans = tabsTransmission(dx,tabs,D)
-        # scale to ADC units
-        return lo + (hi-lo)*trans
-    
-    def fitPhysicalModel(self,samples,tabs,args,direction,lo,hi,offset):
-    
-        # initial parameter guesses
-        loGuess = lo
-        hiGuess = hi
-        t0Guess = offset
-        DGuess = 1.7
-        LGuess = 1020.
-        dthetaGuess = 4.66
-    
-        # initialize model prediction
-        global prediction
-        prediction = numpy.zeros_like(samples)
-    
-        # define chi-square function to use
-        def chiSquare(t0,lo,hi,D,L,dtheta):
-            global prediction
-            prediction = physicalModel(t0,direction,lo,hi,tabs,D,L,dtheta,
-                nsamples=args.nsamples,adcTick=args.adc_tick)
-            residuals = samples - prediction
-            return numpy.dot(residuals,residuals)
-    
-        # initialize fitter
-        print_level = 1 if args.verbose else 0
-        engine = Minuit(chiSquare,errordef=1.0,print_level=print_level,
-            t0=t0Guess,error_t0=1.,
-            lo=loGuess,error_lo=1.,
-            hi=hiGuess,error_hi=1.,
-            D=DGuess,error_D=0.1,fix_D=False,
-            L=LGuess,error_L=10.,fix_L=True,
-            dtheta=dthetaGuess,error_dtheta=0.5,fix_dtheta=False)
-        engine.tol = 10.
-    
-        # do the fit
-        minimum = engine.migrad()
-        if not minimum[0]['has_valid_parameters']:
-            raise RuntimeError('Fit failed!')
-    
-        # calculate the best fit model
-        chiSquare(*engine.args)
-    
-        # return best-fit parameter values and best-fit model prediction
-        return engine.args,prediction
-
-class Template(object):
-
-    def __init__(self,data,timestamp):
-        self.template = scipy.interpolate.UnivariateSpline(data[0],data[1],k=3,s=0.)
-        self.templateTimestamp = timestamp
-
-    def fitTemplateModel(self,frame,args,direction,lo,hi,offset,rise,fall):
-        # initial parameter guesses
-        loGuess = lo
-        rngGuess = hi-lo
-        t0Guess = offset
-        durationGuess = (rise[-1] - fall[0])/2.
-        
-        # initialize a vector of ADC sample countsaf,
-        tadc = numpy.linspace(0.,args.nsamples-1.,args.nsamples)
-    
-        # initialize model prediction
-        global prediction
-        prediction = numpy.zeros_like(frame.samples)
-    
-        # define chi-square function to use
-        def chiSquare(t0,duration,lo,rng):
-            global prediction
-            # convert this frame's ADC timing to s values, correcting for the direction of
-            # travel, the estimate fiducial crossing time, and the stretch factor.
-            svec = direction*(tadc - t0)/duration
-            # evaluate self.template at these s values
-            prediction[:] = 1.
-            mask = numpy.abs(svec) < 1 + args.spline_pad
-            prediction[mask] = self.template(svec[mask])
-            # rescale from [0,1] to [lo,hi]
-            prediction = lo + rng*prediction
-            # evaluate the chi-square
-            residuals = frame.samples - prediction
-            return numpy.dot(residuals,residuals)
-    
-        # initialize fitter
-        print_level = 1 if args.verbose else 0
-        engine = Minuit(chiSquare,errordef=1.0,print_level=print_level,
-            t0=t0Guess,error_t0=1.,
-            duration=durationGuess,error_duration=1.,
-            lo=loGuess,error_lo=1.,
-            rng=rngGuess,error_rng=1.)
-        engine.tol = 1.
-    
-        # do the fit
-        minimum = engine.migrad()
-        if not minimum[0]['has_valid_parameters']:
-            raise RuntimeError('Fit failed!')
-    
-        # return best-fit parameter values and best-fit model prediction
-        return engine.args,prediction
+import frame
+import frameProcessor
+import db
 
 def buildSplineTemplate(frames,args):
     """
@@ -182,8 +28,8 @@ def buildSplineTemplate(frames,args):
     risevec = numpy.empty((nframes,args.nfingers))
     fallvec = numpy.empty((nframes,args.nfingers))
     for i in range(nframes):
-        frame = Frame(samples[i])
-        dirvec[i],lovec[i],hivec[i],t0vec[i],risevec[i],fallvec[i],unusedHeight = frame.quickFit(args)
+        fr = frame.Frame(samples[i])
+        dirvec[i],lovec[i],hivec[i],t0vec[i],risevec[i],fallvec[i],unusedHeight = fr.quickFit(args)
     # calculate the mean lo,hi levels
     lo = numpy.mean(lovec)
     hi = numpy.mean(hivec)
@@ -221,419 +67,6 @@ def buildSplineTemplate(frames,args):
     # average and return the resampled values
     resampled /= nframes
     return numpy.vstack((sgrid,resampled))
-
-class Frame(object):
-    def __init__(self,samples):
-        self.samples = samples
-
-    def padded(self,whalf):
-        return numpy.r_[[self.samples[0]]*whalf,self.samples,[self.samples[-1]]*whalf]
-
-    def kernel(self,wlen):
-        return numpy.ones(wlen,'d')/wlen
-
-    def runningAvg(self,wlen=31):
-        # returns a running-average smoothing of self.samples
-        assert self.samples.size > wlen and wlen % 2 == 1
-        cumulativeSum = numpy.cumsum(numpy.insert(self.padded((wlen-1)/2),0,0))
-        return (cumulativeSum[wlen:] - cumulativeSum[:-wlen])/float(wlen)
-
-    def findNMaxes(self,hist,npeaks=3,sharpThreshold=20,levelThreshold=10):
-        """
-        Given a histogram and the expected number of peaks, returns an array of
-        the approximate center of each of the peaks.
-        """
-        lastE = 0
-        # Define shape (n,2)
-        largest = numpy.zeros((0,2),dtype='int16')
-        # Loop over elements in the differences of the histogram of the smoothed graph
-        count = 0
-        for i,e in enumerate(numpy.diff(hist)):
-            # Looking for a sharp sign change
-            if (e <= 0 and lastE >= 0) or (e >= 0 and lastE <= 0):
-                if lastE-e > sharpThreshold:
-                    # Add potential candidates to an array
-                    largest = numpy.append(largest,[[e-lastE,i]],axis=0)
-                    count = count + 1
-            lastE = e
-        # If greater than n choices, choose the best n
-        if count > npeaks:
-            # Get rid of any indices that are too close to each other
-            nDeleted = 0
-            for index,d in enumerate(numpy.diff(largest[:,1])):
-                if d < levelThreshold:
-                    # Collapse two rows into one
-                    largest = numpy.delete(largest,index+1-nDeleted,axis=0)
-                    nDeleted = nDeleted + 1
-        # Choose indices of best n candidates
-        return largest[numpy.argsort(largest[:,0])][:,1][0:npeaks]
-    
-    def findPeakValues(self,smooth,npeaks=3,window=10):
-        """
-        Given a smoothed array of data with 3 discrete levels (low-mid-hi) returns a sorted
-        tuple low,mid,hi
-        """
-        hist = numpy.histogram(smooth,bins=1023,range=(0,1023))
-        # Average values
-        averages = numpy.zeros(npeaks)
-        i = 0
-        for max in self.findNMaxes(hist[0],npeaks=npeaks):
-            lowerBound = max - window
-            upperBound = max + window
-            # Bounds Checking
-            if lowerBound < 0:
-                lowerBound = 0
-            if upperBound > len(hist[0]):
-                upperBound = len(hist[0])
-            averages[i] = numpy.average(hist[1][lowerBound:upperBound],weights=hist[0][lowerBound:upperBound])
-            i = i+1
-        return tuple(numpy.sort(averages))
-    
-    def findRiseAndFallPositions(self,smooth,lo,mid,hi,nfingers=5,threshold=2):
-        riseMask = numpy.zeros_like(smooth[1:])
-        fallMask = numpy.zeros_like(smooth[1:])
-        # Create boolean Mask where we expect falling edges
-        booleanValue = 1
-        for index,val in enumerate(smooth[1:]):
-            if abs(val-mid) < threshold and booleanValue is 0:
-                booleanValue = 1
-            elif abs(val+mid) < threshold and booleanValue is 1:
-                booleanValue = 0
-            fallMask[index] = booleanValue
-        # Create boolean Mask where we expect rising edges
-        booleanValue = 1
-        for index,val in enumerate(smooth[1:][::-1]):
-            if abs(val-mid) < threshold and booleanValue is 0:
-                booleanValue = 1
-            elif abs(val+mid) < threshold and booleanValue is 1:
-                booleanValue = 0
-            riseMask[-index] = booleanValue
-        # Get Rising Edges
-        rising = numpy.logical_and(numpy.logical_and(smooth[:-1] <= 0,smooth[1:] > 0),riseMask)
-        nrise = numpy.count_nonzero(rising)
-        risePos = numpy.sort(numpy.argsort(rising)[-1*nrise:])
-        # Get Falling Edges
-        falling = numpy.logical_and(numpy.logical_and(smooth[:-1] > 0, smooth[1:] <= 0),fallMask)
-        nfall = numpy.count_nonzero(falling)
-        fallPos = numpy.sort(numpy.argsort(falling)[-1*nfall:])
-        # Raise Exception if there is a problem
-        if nrise != nfingers:
-            raise RuntimeError("findRiseAndFallPositions: expected %d rising edges but found %d" % (nfingers, nrise))
-        if nfall != nfingers:
-            raise RuntimeError("findRiseAndFallPositions: expected %d falling edges but found %d" % (nfingers, nfall))
-        return risePos,fallPos
-    
-    def lineFit(self,y,t1,t2):
-        """
-        Performs a linear fit to determine the value of t where y(t) crosses zero
-        using samples y[t1:t2].
-        """
-        t = numpy.arange(t1,t2)
-        slope, intercept, r_value, p_value, std_err = linregress(t,y[t1:t2])
-        return -intercept/slope
-
-    def getDirection(self,risePos,fallPos,nfingers=5):
-        if risePos[0] - fallPos[0] > self.samples.size/(2.0 * nfingers):
-            direction = +1.
-        else:
-            direction = -1.
-        return direction
-
-    def normalizeWithCenter(self,direction,t0,riseFit,fallFit):
-        if direction == +1:
-            riseFit -= t0
-            fallFit -= t0
-        elif direction == -1:
-            tmp = numpy.copy(riseFit)
-            riseFit = (t0 - fallFit)[::-1]
-            fallFit = (t0 - tmp)[::-1]
-        return riseFit,fallFit
-    
-    def quickFit(self,args,smoothing=15,fitsize=5,avgWindow=50):
-        """
-        Attempts a quick fit of the specified sample data or returns a RuntimeError.
-        Returns the direction (+/-1) of travel, the estimated lo and hi ADC levels,
-        the estimated fiducial crossing time t0 relative to the first sample, and
-        the estimated rise and fall times relative to the fiducial and corrected for
-        the direction of travel. All times are measured in ADC samples.
-        """
-        # perform a running-average smoothing of the frame's raw sample data
-        smooth = self.runningAvg(wlen=1+2*smoothing)
-        # Find the range of the smoothed data. Use the min of the smooth samples to estimate the
-        # lo value. Use the mean of the left and right margins to estimate the hi value. The
-        # reason why don't use the max of the smooth samples to estimate the hi value is that we
-        # observe some peaking (transmission > 1) near the edges.
-        lo,height,hi = self.findPeakValues(smooth)
-        # find edges as points where the smoothed data crosses the midpoints between lo,hi
-        midpt = 0.5*(lo+hi)
-        smooth -= midpt
-        risePos,fallPos = self.findRiseAndFallPositions(smooth,lo,midpt,hi,nfingers=args.nfingers)
-        # perform linear fits to locate each edge to subsample precision
-        riseFit = numpy.empty((args.nfingers,))
-        fallFit = numpy.empty((args.nfingers,))
-        for i in range(args.nfingers):
-            riseFit[i] = self.lineFit(smooth,risePos[i]-fitsize,risePos[i]+fitsize+1)
-            fallFit[i] = self.lineFit(smooth,fallPos[i]-fitsize,fallPos[i]+fitsize+1)
-        # t0 is the time between the 3rd falling and the 3rd rising edge
-        t0 = (fallFit[2]+riseFit[2])/2
-        # use the distance between the first falling and rising edges to 
-        # discriminate between the two possible directions of travel
-        direction = self.getDirection(risePos,fallPos,nfingers=args.nfingers)
-        # calculate edge times relative to the fiducial, corrected for the direction of travel.
-        riseFit,fallFit = self.normalizeWithCenter(direction,t0,riseFit,fallFit)
-        if args.verbose:
-            print direction,lo,hi,t0,riseFit,fallFit,height
-        return direction,lo,hi,t0,riseFit,fallFit,height
-
-class FrameProcessor(object):
-    def __init__(self,tabs,args,db):
-        self.tabs = tabs
-        self.args = args
-        self.db = db
-        self.lastDirection = None
-        self.nextLastOffset = None
-        self.lastOffset = None
-        self.lastElapsed = None
-        self.lastAmplitude = None
-        self.lastSamplesSinceBoot = None
-        self.nextLastSamplesSinceBoot = None
-        self.mostRecentTemplateTimestamp = None
-        # we use python arrays here since we do not know the final size in advance
-        # and they are more time efficient for appending than numpy arrays (but
-        # less space efficient)
-        self.periods = [ ]
-        self.swings = [ ]
-        # initialize plot display if requested
-        if self.args.show_plots:
-            self.fig = plt.figure(figsize=(8,8))
-            plt.ion()
-            plt.show()
-            self.plotx = numpy.arange(self.args.nsamples)
-        # load template if requested
-        if self.args.load_template is not None:
-            if self.args.physical:
-                raise RuntimeError('Options --physical and --load-template cannot be combined')
-            if self.args.load_template == "db":
-                # Load from database
-                templateTuple = self.db.loadTemplate()
-                templateData = templateTuple[0]
-                self.mostRecentTemplateTimestamp = templateTuple[1]
-            else:
-                #load from file specified
-                templateData = numpy.transpose(numpy.loadtxt(self.args.load_template))
-            try:
-                self.template = Template(templateData,self.mostRecentTemplateTimestamp)
-            except IndexError:
-                self.template = None
-                # Output zeroes
-        else:
-            self.template = None
-
-    def updateTemplate(self):
-        """
-        Checks the database for a newer version of the template.  If one exists, replace
-        self.template with it and replace self.mostRecentTemplateTimestamp with its
-        timestamp
-        """
-        if self.args.verbose:
-            print "Attempting to Update Template"
-        if self.args.load_template == "db":
-            # Try and create a template if one doesn't exist
-            if self.template is None:
-                # Try and create one
-                dataTuple = self.db.loadData()
-                data = dataTuple[0]
-                timestamp = dataTuple[1]
-                if len(data.shape) != 1:
-                    if self.args.verbose:
-                        print "Bad data shape!"
-                    return
-                # loop over data frames
-                nframe = len(data)/(1+self.args.nsamples)
-                if(nframe < self.args.fetch_limit):
-                    if self.args.verbose:
-                        print "Insufficient Data to update template!"
-                    return
-                if not (self.args.max_frames == 0 or nframe <= self.args.max_frames):
-                    nframe = self.args.max_frames
-                frames = data[:nframe*(1+self.args.nsamples)].reshape((nframe,1+self.args.nsamples))
-                template = buildSplineTemplate(frames,self.args)
-                # Save to database as array of ordered pairs (arrays)
-                self.db.saveTemplate(template,timestamp)
-            # Look for newer template from database
-            templateTuple = self.db.loadTemplate()
-            templateData = templateTuple[0]
-            templateTimestamp = templateTuple[1]
-            if templateTimestamp != self.mostRecentTemplateTimestamp:
-                self.template = Template(templateData,templateTimestamp)
-                self.mostRecentTemplateTimestamp = templateTimestamp
-
-
-
-    def process(self,samplesSinceBoot,frame):
-        """
-        Processes the next frame of raw IR ADC samples. The parameter samplesSinceBoot gives the number
-        of ADC samples elapsed between the first sample of this frame and the last boot packet recieved, 
-        or zero if this information is not available. Returns the estimated
-        period in seconds (nominally 2 secs) or raises a RuntimeError.
-        """
-        if len(frame.samples) != self.args.nsamples:
-            # Something is seriously wrong.
-            return -2,-2,-2
-        if self.args.load_template and self.template is None:
-            return -3,-3,-3
-        # always start with a quick fit
-        direction,lo,hi,offset,rise,fall,height = frame.quickFit(self.args)
-        if self.args.physical:
-            fitParams,bestFit = fitPhysicalModel(frame.samples,self.tabs,self.args,direction,lo,hi,offset)
-            offset,amplitude = fitParams[0],fitParams[5]
-        elif self.template is not None:
-            fitParams,bestFit = self.template.fitTemplateModel(frame,self.args,
-                direction,lo,hi,offset,rise,fall)
-            offset,duration = fitParams[0],fitParams[1]
-            # convert duration in ADC ticks to a swing amplitude
-            velocity = 0.5*self.args.width/(duration*self.args.adc_tick) # mm/sec
-            if len(self.periods) > 0:
-                period = self.periods[-1]
-            else:
-                period = 2.
-            cosTheta = 1 - 0.5*(velocity*period/(2*math.pi*self.args.length))**2
-            amplitude = math.degrees(math.acos(cosTheta))
-        else:
-            amplitude = None
-            bestFit = None
-        # update our plots, if requested
-        if self.args.show_plots and not self.args.batch_replay:
-            plt.clf()
-            ax = plt.subplot(1,1,1)
-            ax.set_ylim([-4.,1024.])
-            plt.plot(self.plotx,frame.samples,'g+')
-            if self.args.show_centers:
-                plt.plot(numpy.zeros(100)+offset,numpy.linspace(0,2**10,100),'b-');
-            if self.args.show_levels:
-                plt.plot(numpy.linspace(0,self.args.nsamples,100),numpy.zeros(100)+lo,'y-');
-                plt.plot(numpy.linspace(0,self.args.nsamples,100),numpy.zeros(100)+height,'y-');
-                plt.plot(numpy.linspace(0,self.args.nsamples,100),numpy.zeros(100)+hi,'y-');
-            if bestFit is not None:
-                plt.plot(self.plotx,bestFit,'r-')
-            plt.draw()
-        # Calculate the elapsed time since the last dead-center crossing in the same direction,
-        # which is nominally 2 seconds, and the peak-to-peak angular amplitude of two consecutive
-        # swings in opposite directions.
-        if direction != self.lastDirection:
-            if self.nextLastSamplesSinceBoot and self.lastOffset:
-                period = (samplesSinceBoot - self.nextLastSamplesSinceBoot - self.lastOffset + offset)*self.args.adc_tick
-            else:
-                # We must have processed at least two frames before we can calculate a period.
-                period = 0
-            if self.lastAmplitude is not None:
-                swing = amplitude + self.lastAmplitude
-            else:
-                swing = 0
-        else:
-            # We require alternating directions, so something went wrong but we will try to recover.
-            period = -1
-            swing = -1
-        # update our saved info
-        self.lastDirection = direction
-        self.lastOffset = self.nextLastOffset
-        self.nextLastOffset = offset
-        self.nextLastSamplesSinceBoot = self.lastSamplesSinceBoot
-        self.lastSamplesSinceBoot = samplesSinceBoot
-        self.lastAmplitude = amplitude
-        # remember this result
-        if abs(period-2) < 0.1 and swing > 0:
-            self.periods.append(period)
-            self.swings.append(swing)
-        # return the estimated period in seconds
-        return period,swing,height
-
-    def finish(self):
-        periods = numpy.array(self.periods)
-        swings = numpy.array(self.swings)
-        if self.args.show_plots:
-            if self.args.batch_replay:
-                plt.clf()
-                plt.subplot(2,1,1)
-                mean = numpy.mean(periods)
-                # calculate the going in ppm
-                going = 1e6*(periods - mean)/mean
-                plt.plot(going,'x')
-                plt.subplot(2,1,2)
-                plt.plot(swings,'x')
-            self.fig.savefig('fit.png')
-        numpy.savetxt('fit.dat',numpy.vstack([periods,swings]).T)
-
-class DB(object):
-    def __init__(self,args):
-        # Set fetch constants
-        self.ID              = u"_id"
-        self.RAW             = u"raw"
-        self.TIMESTAMP       = u"timestamp"
-        self.CRUDE_PERIOD    = u"crudePeriod"
-        self.TEMPLATE        = u"template"
-
-        self.args = args
-        # Connect to Mongod (fails and exits if mongod is not running)
-        self.client = MongoClient('localhost', args.mongo_port)
-        # Get db object
-        self.db = self.client[args.db_name]
-        # Get collection object
-        self.dataCollection = self.db[args.collection_name]
-        self.templateCollection = self.db[self.args.template_collection]
-
-    def getNumTemplates(self):
-        return self.templateCollection.count()
-
-    # We DONT need the crude period data to generate a template
-    def loadData(self):
-        """
-        Loads IR data from the database into the expected numpy format
-        Crude Period        EDIT 3/27: this is set to zero breaking the --from-db option!
-        IR
-        IR...
-        Returns the tuple (data,timestamp)
-        """
-        # Perform Query
-        results = self.dataCollection.find({},{self.ID:False,self.TIMESTAMP:True, self.CRUDE_PERIOD:True, 
-            self.RAW:True}).sort(self.TIMESTAMP, DESCENDING).limit(self.args.fetch_limit)
-        # Construct numpy array
-        data = numpy.array([], dtype=numpy.uint16)
-        timestamp = 0
-        for document in results:
-            # data = numpy.append(data,document[self.CRUDE_PERIOD])
-            data = numpy.append(data,0)
-            data = numpy.append(data,document[self.RAW])
-            # Get last timestamp
-            timestamp = document[self.TIMESTAMP]
-        return (data,timestamp)
-
-    def saveTemplate(self,template,timestamp):
-        """
-        Saves a given template into the database
-        """
-        self.templateCollection.insert({self.TIMESTAMP:timestamp, self.TEMPLATE:template.T.tolist()})
-        if self.args.verbose:
-            print "Template Updated!"
-        if self.args.show_plots:
-            plt.plot(template[0],template[1])
-            plt.draw()
-
-    def loadTemplate(self):
-        """
-        Loads a given template from the database, returns the tuple (template,timestamp)
-        """
-        # Perform Query
-        results = self.templateCollection.find({},{self.ID:False,self.TIMESTAMP:True, 
-            self.TEMPLATE:True}).sort(self.TIMESTAMP, DESCENDING).limit(1)
-
-        data = []
-        templateTimestamp = None
-        for document in results:
-            templateTimestamp = document[self.TIMESTAMP]
-            for pair in document[self.TEMPLATE]:
-                data.append(pair)
-        return (numpy.transpose(numpy.array(data)),templateTimestamp)
 
 def main():
 
@@ -694,15 +127,15 @@ def main():
     tabs = numpy.array([[-27.,-19.],[-15.,-11.],[-7.,-3.],[3.,7.],[11.,15.],[23.,27.]])
 
     # initialize our database connection
-    db = DB(args)
+    database = db.DB(args)
 
     # initialize our frame processor
-    processor = FrameProcessor(tabs,args,db)
+    processor = frameProcessor.FrameProcessor(tabs,args,database)
 
     # replay a pre-recorded data file if requested
     if args.replay or args.from_db:
         if args.from_db:
-            dataTuple = db.loadData()
+            dataTuple = database.loadData()
             data = dataTuple[0]
             timestamp = dataTuple[1]
         elif args.replay:
@@ -727,7 +160,7 @@ def main():
             template = buildSplineTemplate(frames,args)
             if args.save_template == "db":
                 # Save to database as array of ordered pairs (arrays)
-                db.saveTemplate(template,timestamp)
+                database.saveTemplate(template,timestamp)
             else:
                 # Save to file
                 numpy.savetxt(args.save_template,template.T)
@@ -735,11 +168,10 @@ def main():
                 plt.plot(template[0],template[1])
                 plt.show()
             return 0
-        for i,frame in enumerate(frames):
-            samplesSinceBoot,samples = frame[0],frame[1:]
-            frame = Frame(samples)
+        for i,frame_data in enumerate(frames):
+            samplesSinceBoot,samples = frame_data[0],frame_data[1:]
             try:
-                period,swing,height = processor.process(samplesSinceBoot,Frame(samples))
+                period,swing,height = processor.process(samplesSinceBoot,frame.Frame(samples))
                 print 'Period = %f secs, swing = %f, height = %f (%d/%d)' % (period,swing,height,i,nframe)
                 # check for more recent template
                 processor.updateTemplate()
@@ -766,7 +198,7 @@ def main():
                     samples[args.nsamples - remaining] = value
                     remaining -= 1
                     if remaining == 0:
-                        period,swing,height = processor.process(samplesSinceBoot,Frame(samples))
+                        period,swing,height = processor.process(samplesSinceBoot,frame.Frame(samples))
                         # period,swing,height = 0,0,500
                         # send the calculated period to our STDOUT and flush the buffer!
                         print period,swing,height
